re-organize code for broker messages services (redis, rabbitmq, sqs)
