+++ navigator/brokers/rabbitmq/__init__.py
@@ -1,3 +1,4 @@
 """
 Using RabbitMQ as Message Broker.
 """
+from .connection import RabbitMQConnection
