+++ navigator/brokers/sqs/__init__.py
@@ -3,4 +3,4 @@ AWS SQS Message Broker.
 
 Using Amazon SQS as a Message Broker.
 """
-from .sqs import SQSConnection
+from .connection import SQSConnection
